% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_node.R
\name{ggplot_add.highlight_node}
\alias{ggplot_add.highlight_node}
\title{ggplot_add.highlight_node}
\usage{
\method{ggplot_add}{highlight_node}(object, plot, object_name)
}
\arguments{
\item{object}{An object to add to the plot}

\item{plot}{The ggplot object to add object to}

\item{object_name}{The name of the object to add}
}
\description{
ggplot_add.highlight_node
}
