% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_node.R
\name{highlight_node}
\alias{highlight_node}
\title{highlight_node
highlight the nodes}
\usage{
highlight_node(
  node_name = NULL,
  highlight_color = NULL,
  filter = NULL,
  glow = FALSE,
  glow_fixed_color = FALSE,
  glow_base_size = FALSE,
  highlight_by_shape = FALSE,
  shape_number = NULL,
  shape_size = NULL,
  shape_color = NULL,
  shape_with_text = NULL,
  with_text = NULL,
  glow_size = 1.2,
  override_text = FALSE,
  use_ggfx = NULL,
  ggfx_params = list(),
  textpath_params = list()
)
}
\arguments{
\item{node_name}{used to subset node data with `name` argument}

\item{highlight_color}{highlighting color}

\item{filter}{used to subset node data specified as string}

\item{glow}{use glow highlighting}

\item{glow_fixed_color}{use glow highlighting with fixed color,
specified by `highlight_color` (default: FALSE, use the raw color of the nodes)}

\item{glow_base_size}{use base size of the node for initial size of glowing}

\item{highlight_by_shape}{boolean, use shape highlighting}

\item{shape_number}{shape number}

\item{shape_size}{shape size}

\item{shape_color}{shape color}

\item{shape_with_text}{if specified, show text along with shape (21 only)
using `geomtextpath`}

\item{with_text}{if specified, show text around the nodes, not the shape}

\item{glow_size}{argument to control how big the glowing will be}

\item{override_text}{after highlighting, stack the last geom_node_text layer}

\item{use_ggfx}{use ggfx geom to highlight, default to NULL}

\item{ggfx_params}{ggfx parameters}
}
\description{
highlight_node
highlight the nodes
}
