% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_node_label.R
\name{highlight_node_label}
\alias{highlight_node_label}
\title{highlight_node_label}
\usage{
highlight_node_label(
  node_name = NULL,
  highlight_color = "red",
  highlight_bg_color = "yellow",
  filter = NULL,
  seed = 42
)
}
\arguments{
\item{node_name}{text labels to highlight}

\item{highlight_color}{specify highlighting colour}

\item{highlight_bg_color}{specify highlighting colour}

\item{filter}{text attributes to filter (in string)}

\item{seed}{make repel reproducible}
}
\description{
This just stacks the layer changing the colour of 
specified label in previous layer's output.
Must specify seed to use with ggrepel, and should be 
used with shadowtext in some environments.
}
