% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight_edge.R
\name{highlight_edge}
\alias{highlight_edge}
\title{highlight_edge
highlight edges}
\usage{
highlight_edge(
  edge_name = NULL,
  highlight_color = "red",
  directed = TRUE,
  filter = NULL,
  change_label_color = TRUE,
  glow = FALSE,
  glow_fixed_color = FALSE,
  glow_base_size = FALSE,
  glow_edge_size = 0.3,
  use_ggfx = NULL,
  glow_base_size_manual = 0.01,
  ggfx_params = list()
)
}
\arguments{
\item{edge_name}{edges to highlight
vector of two nodes, or matrix of two nodes
if name argument is not in the graph, index value}

\item{highlight_color}{highlighting color}

\item{directed}{whether to treat `edge_name` as 
directed or undirected, default to TRUE}

\item{filter}{filter arguments of edge attributes in string}

\item{change_label_color}{whether to change label colors}

\item{glow}{use glow highlighting}

\item{glow_fixed_color}{use glow highlighting with fixed color,
specified by `highlight_color` (default: FALSE, use the raw color of the nodes)}

\item{glow_base_size}{use base size of the node for initial size of glowing}

\item{glow_edge_size}{argument to control how big the glowing will be}

\item{use_ggfx}{use ggfx geom to highlight, default to NULL}

\item{glow_base_size_manual}{use this base size if glow_base_size is FALSE}

\item{ggfx_params}{ggfx parameters}
}
\description{
highlight_edge
highlight edges
}
